chore: fix circular deps
