+++ packages/core/remotes/categories/crm/base.ts
@@ -1,25 +1,21 @@
-import { CommonObjectDef, ConnectionUnsafe, CRMProvider, StandardOrCustomObjectDef } from '@supaglue/types';
-import { CRMCommonObjectType, CRMCommonObjectTypeMap, CRMProviderName } from '@supaglue/types/crm';
-import { Association, AssociationCreateParams } from '@supaglue/types/crm/association';
-import { AssociationType, AssociationTypeCreateParams, SGObject } from '@supaglue/types/crm/association_type';
-import { CustomObject, CustomObjectCreateParams, CustomObjectUpdateParams } from '@supaglue/types/crm/custom_object';
-import {
+import type { CommonObjectDef, StandardOrCustomObjectDef } from '@supaglue/types';
+import type { CRMCommonObjectType, CRMCommonObjectTypeMap } from '@supaglue/types/crm';
+import type { Association, AssociationCreateParams } from '@supaglue/types/crm/association';
+import type { AssociationType, AssociationTypeCreateParams, SGObject } from '@supaglue/types/crm/association_type';
+import type {
+  CustomObject,
+  CustomObjectCreateParams,
+  CustomObjectUpdateParams,
+} from '@supaglue/types/crm/custom_object';
+import type {
   CustomObjectRecord,
   CustomObjectRecordCreateParams,
   CustomObjectRecordUpdateParams,
 } from '@supaglue/types/crm/custom_object_record';
-import { FieldMappingConfig } from '@supaglue/types/field_mapping_config';
-import { EventEmitter } from 'events';
+import type { FieldMappingConfig } from '@supaglue/types/field_mapping_config';
 import { Readable } from 'stream';
-import { ConnectorAuthConfig } from '..';
-import { AbstractRemoteClient, RemoteClient } from '../base';
-import * as capsule from '../impl/capsule';
-import * as hubspot from '../impl/hubspot';
-import * as ms_dynamics_365_sales from '../impl/ms_dynamics_365_sales';
-import * as pipedrive from '../impl/pipedrive';
-import * as salesforce from '../impl/salesforce';
-import * as zendesk_sell from '../impl/zendesk_sell';
-import * as zoho_crm from '../impl/zoho_crm';
+import type { RemoteClient } from '../../base';
+import { AbstractRemoteClient } from '../../base';
 
 export interface CrmRemoteClient extends RemoteClient {
   listCommonProperties(object: CommonObjectDef): Promise<string[]>;
@@ -146,59 +142,4 @@ export abstract class AbstractCrmRemoteClient extends AbstractRemoteClient imple
   public async createAssociation(params: AssociationCreateParams): Promise<Association> {
     throw new Error('Not implemented');
   }
-}
-export abstract class CrmRemoteClientEventEmitter extends EventEmitter {}
-
-export type CrmConnectorConfig<T extends CRMProviderName> = {
-  authConfig: ConnectorAuthConfig;
-  newClient: (connection: ConnectionUnsafe<T>, provider: CRMProvider) => AbstractCrmRemoteClient;
-};
-
-export const crmConnectorConfigMap: {
-  [K in CRMProviderName]: CrmConnectorConfig<K>;
-} = {
-  salesforce,
-  hubspot,
-  pipedrive,
-  zendesk_sell,
-  ms_dynamics_365_sales,
-  capsule,
-  zoho_crm,
-};
-
-export function getCrmRemoteClient<T extends CRMProviderName>(
-  connection: ConnectionUnsafe<T>,
-  provider: CRMProvider
-): CrmRemoteClient {
-  const { newClient } = crmConnectorConfigMap[connection.providerName];
-  const client = newClient(connection, provider);
-
-  // Intercept and log errors to remotes
-  return new Proxy(client, {
-    get(target, p) {
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      const v = target[p];
-      if (typeof v !== 'function') {
-        return v;
-      }
-
-      return new Proxy(v, {
-        apply(_target, thisArg, argArray) {
-          try {
-            const res = v.apply(target, argArray);
-            if (Promise.resolve(res) === res) {
-              // if it's a promise
-              return (res as Promise<unknown>).catch((err) => {
-                throw target.handleErr(err);
-              });
-            }
-            return res;
-          } catch (err: unknown) {
-            throw target.handleErr(err);
-          }
-        },
-      });
-    },
-  });
 }
