+++ packages/core/remotes/categories/engagement/base.ts
@@ -1,0 +1,54 @@
+import type { EngagementCommonObjectType, EngagementCommonObjectTypeMap } from '@supaglue/types/engagement';
+import { Readable } from 'stream';
+import { AbstractRemoteClient, RemoteClient } from '../../base';
+
+export interface EngagementRemoteClient extends RemoteClient {
+  listCommonObjectRecords(commonObjectType: EngagementCommonObjectType, updatedAfter?: Date): Promise<Readable>;
+  getCommonObjectRecord<T extends EngagementCommonObjectType>(
+    commonObjectType: T,
+    id: string
+  ): Promise<EngagementCommonObjectTypeMap<T>['object']>;
+  createCommonObjectRecord<T extends EngagementCommonObjectType>(
+    commonObjectType: T,
+    params: EngagementCommonObjectTypeMap<T>['createParams']
+  ): Promise<string>;
+  updateCommonObjectRecord<T extends EngagementCommonObjectType>(
+    commonObjectType: T,
+    params: EngagementCommonObjectTypeMap<T>['updateParams']
+  ): Promise<string>;
+}
+
+export abstract class AbstractEngagementRemoteClient extends AbstractRemoteClient implements EngagementRemoteClient {
+  public constructor(...args: ConstructorParameters<typeof AbstractRemoteClient>) {
+    super(...args);
+  }
+
+  public handleErr(err: unknown): unknown {
+    return err;
+  }
+
+  public async listCommonObjectRecords(
+    commonObjectType: EngagementCommonObjectType,
+    updatedAfter?: Date
+  ): Promise<Readable> {
+    throw new Error('Not implemented');
+  }
+  public async getCommonObjectRecord<T extends EngagementCommonObjectType>(
+    commonObjectType: T,
+    id: string
+  ): Promise<EngagementCommonObjectTypeMap<T>['object']> {
+    throw new Error('Not implemented');
+  }
+  public async createCommonObjectRecord<T extends EngagementCommonObjectType>(
+    commonObjectType: T,
+    params: EngagementCommonObjectTypeMap<T>['createParams']
+  ): Promise<string> {
+    throw new Error('Not implemented');
+  }
+  public async updateCommonObjectRecord<T extends EngagementCommonObjectType>(
+    commonObjectType: T,
+    params: EngagementCommonObjectTypeMap<T>['updateParams']
+  ): Promise<string> {
+    throw new Error('Not implemented');
+  }
+}
