+++ packages/core/remotes/impl/zoho_crm/index.ts
@@ -1,6 +1,6 @@
 import { ConnectionUnsafe, Provider } from '@supaglue/types';
-import { ConnectorAuthConfig } from '../..';
-import { AbstractCrmRemoteClient } from '../../categories/crm';
+import type { ConnectorAuthConfig } from '../../base';
+import { AbstractCrmRemoteClient } from '../../categories/crm/base';
 
 class ZohoCrmClient extends AbstractCrmRemoteClient {
   public constructor() {
